# perf(text): make `levenshteinDistance()` up to 147x faster