I included a link to the paper and added explanatory comments where the implementation differs from that described in the paper. I hope this addresses the understandability problem.

<details><summary>Benchmark code and results</summary>

```text
cpu: Apple M1
runtime: deno 1.45.2+f4b9d85 (aarch64-apple-darwin)

benchmark      time (avg)        iter/s             (min … max)       p75       p99      p995
--------------------------------------------------------------- -----------------------------

group short strings
main          210.58 µs/iter       4,748.9    (195.12 µs … 396 µs) 208.92 µs 264.54 µs 280.29 µs
this PR         6.36 µs/iter     157,287.5     (6.29 µs … 6.46 µs) 6.41 µs 6.46 µs 6.46 µs

summary
  this PR
   33.12x faster than main

group long strings
main              1.8 s/iter           0.6       (1.78 s … 1.85 s) 1.83 s 1.85 s 1.85 s
this PR        12.25 ms/iter          81.6   (12.21 ms … 12.35 ms) 12.26 ms 12.35 ms 12.35 ms

summary
  this PR
   147.05x faster than main
```

```js
import { assertStrictEquals } from "jsr:@std/assert/strict-equals";
import { levenshteinDistance as main } from "jsr:@std/text/levenshtein-distance";
import { levenshteinDistance as thisPR } from "./text/levenshtein_distance.ts";

const { floor, random } = Math;

function randomString(length, charset) {
  let result = "";
  for (let i = 0; i < length; i++) {
    result += charset[floor(random() * charset.length)];
  }
  return result;
}

const alphabet = "abcdefghijklmnopqrstuvwxyz";
for (
  const { group, maxLength } of [
    { group: "short strings", maxLength: 20 },
    { group: "long strings", maxLength: 2000 },
  ]
) {
  const pairs = [];
  for (let i = 0; i < 100; i++) {
    const a = randomString(floor(random() * maxLength), alphabet);
    const b = randomString(floor(random() * maxLength), alphabet);
    assertStrictEquals(thisPR(a, b), main(a, b));
    pairs.push({ a, b });
  }

  Deno.bench("main", { group }, () => {
    for (const { a, b } of pairs) {
      main(a, b);
    }
  });

  Deno.bench("this PR", { group }, () => {
    for (const { a, b } of pairs) {
      thisPR(a, b);
    }
  });
}
```

</details>

Related to #5465.