@0f-0b 2024-07-24
> Are you able to go through the variable names and tweak them to make them a little clearer?

Even if the variable names were changed to be descriptive, it would still require an understanding of the linked paper to see how the algorithm works, so I think it might actually be clearer if the variable names correspond directly to those used in the paper.