@0f-0b 2024-07-24
What the note says is still true so no.