@iuioiua 2024-07-24
The paper says the complexity is O(kn/w).